+++ src/Action.c
@@ -9,7 +9,7 @@
 /* -------------------------------------------------------
  * Action functions
  */
-ListAction * initAction(void (*action)(struct Element *, void *, float), float timing){
+ListAction * initAction(void (*action)(struct Element *, void *, double), double timing){
     ListAction * la = NULL;
     ActionNode * an = NULL;
 
@@ -92,8 +92,8 @@ void freeListAction(ListAction * action){
     }
 }
 
-static int executeOneAction(Action * action, struct Element * e, float time){
-    float ratio;
+static int executeOneAction(Action * action, struct Element * e, double time){
+    double ratio;
     
     // incrementing time
     action->timeSpent += time;
@@ -132,7 +132,7 @@ static void rewindListAction(ListAction * action){
     }
 }
 
-int executeListAction(ListAction * action, struct Element * e, float time){
+int executeListAction(ListAction * action, struct Element * e, double time){
     ActionNode * node;
     int          isFinished = 1;
     
@@ -371,12 +371,12 @@ int delActionToAction(ListAction * action, long long index){
  * Action pre made action functions
  */
 
-void moveByActionFunction(struct Element * e, void * data, float spentTime){
-    float * infos = (float*)data;
-    float   moveX;
-    float   moveY;
-    float   x;
-    float   y;
+void moveByActionFunction(struct Element * e, void * data, double spentTime){
+    double * infos = (double*)data;
+    double   moveX;
+    double   moveY;
+    double   x;
+    double   y;
     
     if(e && data){
         moveX = infos[0];
@@ -398,12 +398,12 @@ void moveByActionFunction(struct Element * e, void * data, float spentTime){
     }
 }
 
-void moveToActionFunction(struct Element * e, void * data, float spentTime){
-    float * infos = (float*)data;
-    float   newX;
-    float   newY;
-    float   oldX;
-    float   oldY;
+void moveToActionFunction(struct Element * e, void * data, double spentTime){
+    double * infos = (double*)data;
+    double   newX;
+    double   newY;
+    double   oldX;
+    double   oldY;
     
     if(e && data){
         newX  = infos[0];
@@ -425,12 +425,12 @@ void moveToActionFunction(struct Element * e, void * data, float spentTime){
     }
 }
 
-void scaleByActionFunction(struct Element * e, void * data, float spentTime){
-    float * infos = (float*)data;
-    float   x;
-    float   y;
-    float   w;
-    float   h;
+void scaleByActionFunction(struct Element * e, void * data, double spentTime){
+    double * infos = (double*)data;
+    double   x;
+    double   y;
+    double   w;
+    double   h;
     
     if(e && data){
         x     = infos[0];
@@ -452,12 +452,12 @@ void scaleByActionFunction(struct Element * e, void * data, float spentTime){
     }
 }
 
-void scaleToActionFunction(struct Element * e, void * data, float spentTime){
-    float * infos = (float*)data;
-    float   newW;
-    float   newH;
-    float   oldW;
-    float   oldH;
+void scaleToActionFunction(struct Element * e, void * data, double spentTime){
+    double * infos = (double*)data;
+    double   newW;
+    double   newH;
+    double   oldW;
+    double   oldH;
     
     if(e && data){
         newW  = infos[0];
@@ -479,10 +479,10 @@ void scaleToActionFunction(struct Element * e, void * data, float spentTime){
     }
 }
 
-void rotateByActionFunction(struct Element * e, void * data, float spentTime){
-    float * infos = (float*)data;
-    float   toAdd;
-    float   angle;
+void rotateByActionFunction(struct Element * e, void * data, double spentTime){
+    double * infos = (double*)data;
+    double   toAdd;
+    double   angle;
     
     if(e && data){
         toAdd = infos[0];
@@ -496,10 +496,10 @@ void rotateByActionFunction(struct Element * e, void * data, float spentTime){
     }
 }
 
-void rotateToActionFunction(struct Element * e, void * data, float spentTime){
-    float * infos = (float*)data;
-    float   newAngle;
-    float   oldAngle;
+void rotateToActionFunction(struct Element * e, void * data, double spentTime){
+    double * infos = (double*)data;
+    double   newAngle;
+    double   oldAngle;
     
     if(e && data){
         newAngle = infos[0];
@@ -513,8 +513,8 @@ void rotateToActionFunction(struct Element * e, void * data, float spentTime){
         setAngleElement(e, oldAngle + (newAngle - oldAngle) * spentTime);
     }
 }
-void fadeInActionFunction(struct Element * e, void * data, float spentTime){
-    float * infos = (float*)data;
+void fadeInActionFunction(struct Element * e, void * data, double spentTime){
+    double * infos = (double*)data;
     int     oldAlpha;
     int     newAlpha;
     
@@ -523,7 +523,7 @@ void fadeInActionFunction(struct Element * e, void * data, float spentTime){
 
         if(infos[1] < 0){
             getAlphaElement(e, &oldAlpha);
-            infos[1] = (float)oldAlpha;
+            infos[1] = (double)oldAlpha;
         }else{        
             oldAlpha = infos[1];
         }
@@ -531,8 +531,8 @@ void fadeInActionFunction(struct Element * e, void * data, float spentTime){
         setAlphaElement(e, oldAlpha - newAlpha * spentTime);
     }
 }
-void fadeOutActionFunction(struct Element * e, void * data, float spentTime){
-    float * infos = (float*)data;
+void fadeOutActionFunction(struct Element * e, void * data, double spentTime){
+    double * infos = (double*)data;
     int     oldAlpha;
     int     newAlpha;
     
@@ -541,7 +541,7 @@ void fadeOutActionFunction(struct Element * e, void * data, float spentTime){
 
         if(infos[1] < 0){
             getAlphaElement(e, &oldAlpha);
-            infos[1] = (float)oldAlpha;
+            infos[1] = (double)oldAlpha;
         }else{        
             oldAlpha = infos[1];
         }
@@ -558,9 +558,9 @@ void fadeOutActionFunction(struct Element * e, void * data, float spentTime){
 /* -------------------------------------------------------
  * Action pre made actions
  */
-ListAction * moveByAction(float x, float y, float time){
+ListAction * moveByAction(double x, double y, double time){
     ListAction * result = NULL;
-    float      * data   = malloc(4 * sizeof(*data));
+    double      * data   = malloc(4 * sizeof(*data));
 
     if(data){
         data[0] = x;
@@ -578,9 +578,9 @@ ListAction * moveByAction(float x, float y, float time){
     return result;
 }
 
-ListAction * moveToAction(float x, float y, float time){
+ListAction * moveToAction(double x, double y, double time){
     ListAction * result = NULL;
-    float      * data   = malloc(4 * sizeof(*data));
+    double      * data   = malloc(4 * sizeof(*data));
 
     if(data){
         data[0] = x;
@@ -598,9 +598,9 @@ ListAction * moveToAction(float x, float y, float time){
     return result;
 }
 
-ListAction * scaleByAction(float x, float y, float time){
+ListAction * scaleByAction(double x, double y, double time){
     ListAction * result = NULL;
-    float      * data   = malloc(4 * sizeof(*data));
+    double      * data   = malloc(4 * sizeof(*data));
 
     if(data){
         data[0] = x;
@@ -618,9 +618,9 @@ ListAction * scaleByAction(float x, float y, float time){
     return result;
 }
 
-ListAction * scaleToAction(float w, float h, float time){
+ListAction * scaleToAction(double w, double h, double time){
     ListAction * result = NULL;
-    float      * data   = malloc(4 * sizeof(*data));
+    double      * data   = malloc(4 * sizeof(*data));
 
     if(data){
         data[0] = w;
@@ -638,9 +638,9 @@ ListAction * scaleToAction(float w, float h, float time){
     return result;
 }
 
-ListAction * rotateByAction(float angle, float time){
+ListAction * rotateByAction(double angle, double time){
     ListAction * result = NULL;
-    float      * data   = malloc(2 * sizeof(*data));
+    double      * data   = malloc(2 * sizeof(*data));
 
     if(data){
         data[0] = angle;
@@ -656,9 +656,9 @@ ListAction * rotateByAction(float angle, float time){
     return result;
 }
 
-ListAction * rotateToAction(float angle, float time){
+ListAction * rotateToAction(double angle, double time){
     ListAction * result = NULL;
-    float      * data   = malloc(2 * sizeof(*data));
+    double      * data   = malloc(2 * sizeof(*data));
 
     if(data){
         data[0] = angle;
@@ -674,12 +674,12 @@ ListAction * rotateToAction(float angle, float time){
     return result;
 }
 
-ListAction * fadeInAction(int alpha, float time){
+ListAction * fadeInAction(int alpha, double time){
     ListAction * result = NULL;
-    float      * data   = malloc(3 * sizeof(*data));
+    double      * data   = malloc(3 * sizeof(*data));
 
     if(data){
-        data[0] = (float)alpha;
+        data[0] = (double)alpha;
         data[1] = -1;
             
         result = setDataAction(initAction(fadeInActionFunction, time), data, 1);
@@ -692,12 +692,12 @@ ListAction * fadeInAction(int alpha, float time){
     return result;
 }
 
-ListAction * fadeOutAction(int alpha, float time){
+ListAction * fadeOutAction(int alpha, double time){
     ListAction * result = NULL;
-    float      * data   = malloc(3 * sizeof(*data));
+    double      * data   = malloc(3 * sizeof(*data));
 
     if(data){
-        data[0] = (float)alpha;
+        data[0] = (double)alpha;
         data[1] = -1;
             
         result = setDataAction(initAction(fadeOutActionFunction, time), data, 1);
